requirements.txt
async_property==0.2.2
discord_emojis==2023.12.23
hikari==2.0.0.dev126
hikari_miru==4.1.1
hikari-lightbulb==3.0.0a8